Refactoring to make different mapping functions easier to find
 - Fixing Any signature to no longer require a target
